bippy: handle branches that are vulnerable but there is no fix!
We can now create pairs of "vunerable but no fix" branches, and generate
the mbox lines, but the json generation isn't working just yet.

So comment out the generation logic to wait for the CVE group to get
back to me, and keep the logic for finding the pairs.  The generation
logic will be fixed up later.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>