bippy: add the use of dyad to replace logic parsing the kernel git tree commits
Added dyad, a replacement tool for the "find the pairs of problem
commits" logic that is currently in bippy.  The logic in bippy has some
bugs that was getting harder to work around, so start from scratch with
dyad, just for this portion of the overall logic that bippy was
providing.

This can allow us to replace dyad with anything else in the future (rust
all the things!), as well as properly generate the intermediate data in
a format that others can consume (instead of having to read cve json
files) as has been asked for.

dyad fixes a number of bugs in bippy's parsing logic, and has a big
regression test suite (tests, amazing!) in the repo that it is developed
in at:
	https://git.sr.ht/~gregkh/bippy

These fixed issues include, but are not limited to:
  - properly listing branches that are still vulnerable to an issue
  - properly handling multiple "Fixes:" tags in a sane way (i.e. in
    checked-in-commit order, instead of random order)
  - properly handling non-git-sha information in "Fixes:" tags
  - properly finding git commit ids in stable branches for backported
    fixes (the real id, not just a commit chance for where the related
    commit was discussed)
  - properly noticing when a vulnerability is introduced and then fixed
    in the same release and providing that information correctly.

dyad also has the benefit of being a "now that we know what we need to
do, let's start from scratch and do it right" replacement, making for
cleaner code that is hopefully documented better than bippy was, which
evolved as we learned the complexities of just how messy git commits for
the kernel really are (see some of the issues fixed above for things we
have learned.)  Yes, it's still in bash, but as it's just part of the
overall pipeline for creating cve entries, there's no reason it could
not be rewritten into something else if needed/wanted.

Right now the old parsing logic in bippy still generates the cve json
files, but dyad is used to create the mbox information, as mbox
information only gets posted once, while the cve json data can be
updated at any time.

Switching over to dyad now allows all future mail messages to contain
the proper information which is important for valid reports.  Downside
is that running bippy is now twice as long as the data is generated
twice, but cpu time is "free", right?

Work will be done to move the cve json data over to use the output of
dyad, but that requires some adjustments for how the json data is
created and work with the CVE group in how to properly represent some of
this (multiple json lists showing the git information seem to break the
schema, or the cve backend parsing logic, still yet to be determined
which is not correct, odds are it's on our end somewhere)

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>