cve: fix up a bunch of json entries
Turned out we had a bug in the json creations, we would not properly
describe the "affected" kernel version if there were stable kernel
branches older than the mainline kernel that was affected.

This bug is now fixed in bippy, so update all the relevent json entries
that were incorrect.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>