cve: update the .json records with missed git ids
For some branches, the commit is vulnerable, and fixed, in the same
release, but we were not referencing it in the list of git ids in the
json record.  For stable branches, this isn't a big deal, but for many
commits, it is in Linus's release branch, so this is important to list.

This updates all current cve records with the missed information.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>