bippy: update to latest version, removing old fixes logic entirely
Now we are fully relying on dyad for creating the vulnerable:fixes
lines, so remove all of the old logic in bippy that was parsing git logs
and the like.  We are back to the original speed of all of this, before
dyad was added as we are not duplicating the logic in two different
ways.

Testing passes all existing CVE entries to generate the same logic as
before.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>