scripts/cve_publish_json: put check back in
I accidentally checked my the "comment the check out" change, as when
submitting lots of json entries, it takes a long time for the round-trip
to the cve server for the FULL list each time.

Put the check back, ideally this should be changed to only check the
individual commit, which will make things go faster, especially as the
number of ids we have allocated will only keep increasing.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>