dyad: provide fix for when a "Fixes:" line ends up null
Many commits have problems where a Fixes: tag is not correct, or the
matching logic gets it wrong as it's in text form.  When that happened,
we would get a error message from 'git rev-list' splatted out to the
screen, not looking good, but still working properly.

Fix this up by handling an empty list better.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>