scripts/voting_results: rename additional reviewer to "Guest"
As we have others providing results, remove the hardcoded name for now
and rename it as "Guest" for the output and the script variables.  Makes
it easier to provide results "built in" for others as we then only need
to manually change 2 lines of bash script, not lots of others.

Ideally we could add a command line argument, but baby steps for now...

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>