dyad: change to use verhaal's id_find_active script
We have a problem when commits are reverted.  Or added and then
reverted.  Or reverted and added back, or some combination of all of
this.  To handle this, verhaal now can track what reverts what, so use
that information to properly handle branches and commits that are
reverted.

This will fix a number of old cve entries where commits were reverted
and we were inproperly handling this.  Thanks to Rolf for pointing this
out and proding me to finally resolve it.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>