update cve ids based on handling reverts properly now.
Thanks to Eike for pointing out the problem that we had where we did not
properly handle a revert as "fixing" the commit it was reverting if it
did not have an explicit "Fixes:" tag in the changelog body.  This
resulted in 1 cve being made invalid, and 2 more having updates to their
vulnerable ranges.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>