add .vulnerable id for CVE-2022-49616
This commit is not necessarily where the issues were actually introduced. This
commit is where the rt711 codec driver was first introduced into the mainline
kernel, so this is just a lower barrier to denote that no older versions can be
affected. Also affected are the rt700 and RT711 SDCA driver, but they were
introduced later.

Signed-off-by: Rolf Eike Beer <eb@emlix.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>