add .vulnerable id for CVE-2025-21780
This commit is not necessarily where the issues were actually introduced. This
commit is where the support for the new smu type in the AMD gpu driver was
first introduced into the mainline kernel, so this is just a lower barrier to
denote that no older versions can be affected.

Signed-off-by: Rolf Eike Beer <eb@emlix.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>