dyad: properly sort the fixed kernel list
The database isn't in sorted order, so when we read from it, we need to
sort the kernels found so that the output is always in a deterministic
way.

Add a test to check that we can sort a vector of Kernels properly by
version number to verify this keeps working properly.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>