dyad: add test for invalid fixes
Sometimes the db picks up commits that are NOT valid commit ids (I'm
looking at you syzbot message ids...) so we need to properly filter them
out in the fixes list so that we don't get totally confused.

Add a test for this so that we don't break this functionality.
Note, we currently do not pass this test, will fix that up next.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>