dyad: properly handle invalid Fixes: lines
The database can lie about a valid Fixes lines as it does no validation
on if it is correct or not.  So verify each kernel commit id that the
database gives us.

Bonus is we don't need to shell out to bash/git/tac anymore, speeding up
things even more!

Majority of the diff is moving the loop back an indent due to not having
to build up a list of kernels "yet again", just do it once when we first
ask the database for it.

All tests now pass properly.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>