dyad: do NOT fall back to sort -V
We were doing a dumb "sort -V" thing for git ids, with the version at
the front, but that was because we didn't think about versions being
identical.  If they are, then we are sorting by the order the SHA1 comes
in the alphabet, NOT the order in the git tree.  We know how to properly
sort a git list of ids, so use that instead!

Fix up the tests where we had gotten this wrong before, old dyad is not
correct here.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>