dyad: switch to rust version!
dyad is now a symlink to the target release of the rust build.  Also
increment the version number of the Cargo.toml release to be "1.0.0" as
it's ready for real-world use!

Thanks so much to Sasha for all the work put into this, result is much
faster and much more fun to work on.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>