dyad: allow Kernel::new() to fail
Well, it doesn't quite fail yet, but now all callers have to check to
verify if it could fail or not.  This is a step toward making it
"simpler" to create a kernel object by just providing a git id, not both
the git id and version, and allow the kernel object to dig up that
version if/when it is actually needed.  This should happen as right now
it is possible to create a kernel object with an invalid git id, which
is not a good idea.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>