bippy: add framework for cpeApplicability json entries
We want to support cpe ranges, so that groups like NIST doesn't have to
manually enter them in "by hand" and it gives people a simpler way to
figure out vulnerable ranges.  Add the initial json structures to bippy
so that we can build on top of them going forward.

The code does not change any json records yet at all, it's only the data
structures being added at this point in time.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>