tools/cve_utils: fix up comment for empty_kernel()
The comment is exactly opposite for what is happening here, so fix that
up, and add some more tests to verify this is all working properly.

Also, I think the logic might not be necessary anymore, switching it
around didn't seem to break dyad's tests, but more work needs to be
figured out if that's really true or not, so just leave it as-is for
now.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>