tools: dyad: make get_revert() return a Kernel
This cleans up loads of logic in dyad, now we are dealing with a real
kernel object and don't have to go through an intermediate step to take
a git id, save it, and then attempt to create it again.

Logic is cleaner and hopefully easier to follow overall for creation of
vulnerable kernels.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>