tools: bippy: use Kernel object for dyad output
bippy should not be "hand rolling" it's own kernel object, so start to
convert it over to using Kernel (ideally it could just use KernelPair,
but for now let's take baby steps.)  Move to use Kernel for the
"vulnerable" portion of the dyad kernel pairs.

By doing this, lots more validation of if this is a valid kernel object
or not is now done.  The majority of the "churn" in this change is
moving the bippy test suite to actually have real git commit ids instead
of "fake" ones.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>