tools: dyad: allow a valid mainline commit to work that is not in a released kernel
Sometimes we have git ids that are backported BEFORE they hit a released
Linus kernel.  The rust version of dyad would fail on that, so fix that
up by doing additional checks to see if this is actually a valid id in
the repo, but just not assigned to a version yet.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>