tools: bippy: report git ranges that are NOT fixed in the JSON output
The original reporting does NOT report where we have known vulnerable
git ranges in the json output.  This gives users the misguided notion
that those ranges/branches ARE fixed, when specifically they ARE NOT
fixed at all.

Originally this was done because users did not seem to be able to parse
git ranges well, but due to users totally ignoring many vulnerable
kernel branches, this was deemed to "not be a good thing", so let's
properly report this information now as we do know it.

Also, with the addition of the CPE information to the json output, this
is even more obvious where ranges are not fixed, and that's also visable
in the version ranges, so to ignore that when reporting the git
information is not a good thing at all.

Fix this up and now report git ranges that are not fixed.  Let's see if
people strongly object, if so, this commit can be reverted and the json
records regenerated and updated, but hopefully it does not come to that.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>