tools: bippy: remove some redundant args parsing
bippy should not need any additional arguments passed to it, so if it
gets them, abort and print out what went wrong, don't attempt to treat
them as extra "references", that should only be in the --references
argument passed to the tool.

Removes a bunch of code that turned out never to be called, always a
good thing.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>