tools: bippy: do NOT assign a CVE if we do not have any vulnerable kernels
We are not allowed to create a CVE for an id that is vulnerable:fixed in
the same kernel release, so if we have NO vulnerable:fixed ranges that
are not in the same release, we have to abort and not create it.

This matches the "old" bippy logic, somehow this snuck in with the
rewrite to Rust.  Now a few CVEs that were allocated but should not have
been, need to go be revoked...

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>