tools: bippy: do not create a json file if the mbox will not work
If we do not have a valid set of vulnerable:fixed pairs that span a
release, then the mbox file will not be created, BUT the json file could
have been, which causes confusion.  So always create the mbox file
first to catch this.

Note, we will still have a problem if a mbox file is not asked to be
created, but for now, that doesn't seem to be anything that we actually
do, so it's ok, but we should revisit this someday...

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>