tools: bippy: move policy decision to not create a CVE id to a more prominent place
bippy should NOT be creating a CVE if there is no vulnerable:fixed
kernel versions.  That logic is buried way down deep in the creation of
the mbox file, which isn't exactly obvious.  Move it higher up to the
main processing logic and document it really well with a ***POLICY***
comment header.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>