tools: bippy: only create the list of affected files once
The call to get_affected_files() is almost ALL of the cpu time in bippy
right now.  Calling it twice is crazy, no need to do that, so only call
it once, halving overall the time it takes to run the program.

So not only did we speed up the code 100%, we removed a bunch of uneeded
and duplicate logic as well, making the whole thing even smaller and
simpler!

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>