tools: dyad: add potential test for fixes older than vulnerable
For many entries, we have fixes in kernels that are older than
vulnerable locations.  Right now we report that, but it doesn't always
make sense to do so.  Here's some code that can be commented out to have
that checked and if it happens, provide the information in a comment
only.

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>