CVE-2021-46926: Add .vulnerable file
The root cause of the issue fixed by commit
385f287f9853da402d94278e59f594501c1d1dad was introduced by commit
6f11586f4896ee448262747788a0a3faf0fe9066 ("soundwire: intel: filter
SoundWire controller device search") authored by Pierre-Louis Bossart on
May 22, 2019.

In commit 6f11586f4896e, the sdw_intel_acpi_cb callback function was
modified to filter for SoundWire controllers. The bug was that the code
set res->handle = handle before checking if the device was actually a
SoundWire controller. This meant that during the ACPI namespace walk:

1. The callback was called for each child device under the HDAudio
   controller
2. For each device, it set res->handle = handle
3. Only after setting the handle did it check if the device was a
   SoundWire controller (by checking bits 31..28 of _ADR)
4. If it wasn't a SoundWire controller, it returned AE_OK to continue
   walking
5. If it was a SoundWire controller, it returned AE_CTRL_TERMINATE to
   stop

The problem scenario: If there were multiple children under the HDAudio
controller and the SoundWire controller was not the last one in the walk
order, then:
- The handle got set to a non-SoundWire device first
- When the actual SoundWire controller was found, the walk terminated
- But res->handle still pointed to the wrong (non-SoundWire) device

This bug could cause the Intel SoundWire driver to:
1. Use the wrong ACPI device handle for the SoundWire controller
2. Potentially access incorrect device resources
3. Fail to properly initialize the SoundWire subsystem
4. In worst cases, cause system crashes when accessing invalid memory
   through the wrong handle

The bug persisted for approximately 2.5 years (May 2019 to December
2021) through multiple code reorganizations before being properly
identified and fixed.

Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>