CVE-2021-47082: Add .vulnerable file
The root cause of the double-free bug fixed by commit
158b515f703e75e7d68289bf4d98c664e1d632df was introduced by commit
608b9977260f67d8b032ea170666a6174a48e2f1 ("tun: use per cpu variables
for stats accounting") on April 13, 2016.

The chain of events:

1. April 13, 2016 - Commit 608b9977260f67 introduced the problematic
   pattern:
   - Added tun->pcpu_stats = netdev_alloc_pcpu_stats() allocation in
     tun_set_iff()
   - Added cleanup in the error path: free_percpu(tun->pcpu_stats)
   - Also added the same cleanup in the destructor tun_free_netdev()
   - At this time, this wasn't a bug because register_netdevice()
     didn't call the destructor on failure

2. May 8, 2017 - Commit cf124db566e6b ("net: Fix inconsistent teardown
   and release of private netdev state") changed the behavior:
   - Modified register_netdevice() to call priv_destructor on failure
   - This exposed the latent double-free bug in the tun driver
   - Now when register_netdevice() failed, it would call
     tun_free_netdev() which would free pcpu_stats
   - Then the error path in tun_set_iff() would try to free the same
     pcpu_stats again

3. Additional resources affected:
   - The same pattern existed for tun->security (added by commit
     5dbbaf2de89613d19a9286d4db0a535ca2735d26 in January 2013)
   - When commit 497a5757ce4e8f changed pcpu_stats to dev->tstats in
     November 2020, the bug persisted with the new field

Commit 608b9977260f67d8b032ea170666a6174a48e2f1 is the commit that
originally introduced the root cause. It created a situation where the
same resources were freed in two places:
1. In the destructor (tun_free_netdev)
2. In the error path of tun_set_iff

This wasn't immediately a problem because the destructor wasn't called
on register_netdevice() failure. However, this created a latent bug that
was triggered over a year later when the kernel's netdevice registration
behavior changed.

The fix in commit 158b515f703e75e7d68289bf4d98c664e1d632df properly
resolved this by moving the allocation to the ndo_init callback, which
is properly paired with the destructor, ensuring resources are only
freed once regardless of where the failure occurs.

Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>